Moving some files, fixing a few bugs
