--- src/loading.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default class Loading extends React.Component {
-  render() {
-    return <div>Loading...</div>
-  }
-}
